+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## Unreleased
+
+- Added `outDrawColor()` and `outFillColor()` functions to 
+  the `PdfDocument` class.
+
 ## 3.0.8 - 2025-02-26
 
 - **BC Break!**: Sets `PdfBorder` properties public, removed the corresponding